+++ GeneradoresYPruebas.Api/Helpers/DatosPruebaFrecuencia.cs
@@ -4,9 +4,9 @@ public class DatosPruebaFrecuencia
 {
     public double Comparador {  get; set; }
     public double[] ValoresU {  get; set; }
-    public int X { get; set; }
+    public double X { get; set; }
 
-    public DatosPruebaFrecuencia(double comparador, double[] valoresU, int x)
+    public DatosPruebaFrecuencia(double comparador, double[] valoresU, double x)
     {
         Comparador = comparador;
         ValoresU = valoresU;
