+++ Pruebas.Tests/FrecuenciaTests.cs
@@ -9,7 +9,7 @@ public class FrecuenciaTests
     public void PruebaPowerpoint()
     {
         double[] valoresU = [0.01, 0.079, 0.168, 0.858, 0.901, 0.74, 0.713, 0.478, 0.277, 0.019, 0.548, 0.426];
-        int x = 3;
+        double x = 3;
         var comparador = 0.65;
 
         var resultado = Frecuencia.EsAleatorio(x, comparador, valoresU);
